+++ lib/baby_squeel.rb
@@ -1,6 +1,5 @@
 require 'active_record'
 require 'active_record/relation'
-# Loading polyamorous from 'ransack' as of commit c9cc20de9 (post v2.3.2)
 require 'polyamorous/polyamorous'
 require 'baby_squeel/version'
 require 'baby_squeel/errors'