+++ spec/spec_helper.rb
@@ -11,6 +11,7 @@ require 'support/models'
 require 'support/matchers'
 require 'support/factories'
 require 'support/query_tracker'
+require 'support/polyamorous_helper'
 
 if ActiveSupport.respond_to?(:deprecator)
   ActiveSupport.deprecator.behavior = :raise
@@ -22,6 +23,8 @@ end
 RSpec.configure do |config|
   config.include Factories
   config.include QueryTracker
+  config.include PolyamorousHelper, :polyamorous
+
   config.filter_run focus: true
   config.default_formatter = config.files_to_run.one? ? :doc : :progress
   config.run_all_when_everything_filtered = true