+++ spec/support/models.rb
@@ -19,6 +19,7 @@ class Post < ActiveRecord::Base
   has_many :pictures, as: :imageable
 
   belongs_to :parent, class_name: 'Post'
+  belongs_to :child, class_name: 'Post'
 end
 
 class Comment < ActiveRecord::Base